requirements.txt
pycryptodome
pwntools
pyjwt
censys
cryptography
z3-solver
requests
